cleanup functional tests
